app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "DoraemonTCat", email = "peeframesave@gmail.com"}
]
requires-python = ">=3.11"   # ← แก้ตรงนี้
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "line-bot-sdk (>=3.21.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "chromadb (>=1.3.5,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
